requirements/common.txt
boto3==1.26.152
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.51.0
django-cors-headers==4.0.0
django-storages==1.13.2
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
pillow==9.5.0
openai==0.27.8